requirements.txt
PyGithub==2.9.0
openai>=1.30.0
pandas>=2.0.0
xlsxwriter>=3.1.0
numpy>=1.24.0